+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - MT-32 ROM/SoundFont switch messages are now only shown when the appropriate synth is active.
+- A message is now displayed on the LCD for a second just before entering power saving mode.
 
 ### Fixed